src/backend/requirements.txt
# Core Framework
fastapi==0.135.2
uvicorn==0.42.0
python-dotenv==0.21.0
pydantic==2.10.6
pydantic-settings==2.7.1

# AWS & Bedrock
boto3==1.42.77
botocore==1.42.77
langchain-aws==0.2.14
langchain-core==0.3.29

# Search
serpapi==1.0.2

# Document Generation
fpdf2==2.8.2
python-docx==1.1.2

# Database (DynamoDB)
# boto3 already included above

# CORS & HTTP
python-multipart==0.0.22
